requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.3.2
pyswisseph==2.10.03
geopy==2.4.1
Flask-CORS==4.0.0 